camada de service totalmente concluída, todos os métodos implementados e agora temos tratamento de exceções

+++ back-end/questoes-java/src/main/java/com/lipe/questoes_br/Repository/PartidaRepository.java
@@ -1,9 +1,8 @@
 package com.lipe.questoes_br.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.lipe.questoes_br.Entity.Partida;
 
-public interface PartidaRepository extends JpaRepository<Partida, Integer> {
+public interface PartidaRepository extends JpaRepository<Partida, Long> {
 
 }
